Restructure files and replace user schema
