+++ frontend/src/components/Login.jsx
@@ -1,5 +1,3 @@
-import './App.css'
-
 function Login() {
     return <h1>login header</h1>
 }
